# static_files/projects/kidk/kidk_openapi.yaml
openapi: 3.0.3
info:
  title: Kidk API
  version: 0.1.0
  description: >
    Initial API contract for Kidk mobile platform.
    Includes authentication, child management, fever monitoring,
    jaundice assessment, consultation, growth and vaccination.
servers:
  - url: https://api.kidk.example.com/v1
security:
  - bearerAuth: []
paths:
  /auth/request-otp:
    post:
      tags: [Auth]
      summary: Request OTP for signup/login/password reset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpRequestInput'
      responses:
        '202':
          description: OTP scheduled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericMessage'
  /auth/signup:
    post:
      tags: [Auth]
      summary: Sign up parent user with verified OTP
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupInput'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with mobile and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
  /auth/change-password:
    post:
      tags: [Auth]
      summary: Change current user password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [oldPassword, newPassword]
              properties:
                oldPassword:
                  type: string
                newPassword:
                  type: string
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericMessage'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    GenericMessage:
      type: object
      properties:
        message:
          type: string
    OtpRequestInput:
      type: object
      required: [mobile, purpose]
      properties:
        mobile:
          type: string
        purpose:
          type: string
          enum: [signup, login, password_reset]
    SignupInput:
      type: object
      required: [mobile, otpCode, firstName, lastName, password]
      properties:
        mobile:
          type: string
        otpCode:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        password:
          type: string
    LoginInput:
      type: object
      required: [mobile, password]
      properties:
        mobile:
          type: string
        password:
          type: string
    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/UserInfo'
    UserInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        mobile:
          type: string
...